# src/documentation/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IProduct": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"stock": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"price",
					"stock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PagedList_IProduct_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/IProduct"
						},
						"type": "array"
					},
					"totalRecords": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"offset": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalRecords",
					"currentPage",
					"offset"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"ApiResponse_PagedList_IProduct__": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"validationErrors": {
						"$ref": "#/components/schemas/FieldErrors"
					},
					"data": {
						"$ref": "#/components/schemas/PagedList_IProduct_"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_IProduct-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"validationErrors": {
						"$ref": "#/components/schemas/FieldErrors"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IProduct"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_IProduct_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"validationErrors": {
						"$ref": "#/components/schemas/FieldErrors"
					},
					"data": {
						"$ref": "#/components/schemas/IProduct"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddProductRequest": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"stock": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"price",
					"stock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseApiResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean"
					},
					"validationErrors": {
						"$ref": "#/components/schemas/FieldErrors"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductRequest": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"stock": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"price",
					"stock",
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "foxentry-petr-lisak",
		"version": "1.0.0",
		"description": "Rozhraní pro pomoc se správou obchůdku panu Petru Lišákovi",
		"contact": {
			"name": "Martin Přívozník"
		}
	},
	"paths": {
		"/api/products": {
			"get": {
				"operationId": "GetProductsPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PagedList_IProduct__"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/products/all": {
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_IProduct-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/products/search": {
			"get": {
				"operationId": "FulltextSearchProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_IProduct-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/products/searchByStockQuantity": {
			"get": {
				"operationId": "SearchProductsByStockQuantity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_IProduct-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "min",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "max",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/products/add": {
			"post": {
				"operationId": "AddProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_IProduct_"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddProductRequest"
							}
						}
					}
				}
			}
		},
		"/api/products/update": {
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductRequest"
							}
						}
					}
				}
			}
		},
		"/api/products/delete/{id}": {
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}